Implement ColorSelectionActivity.onTouch, first pass
When we draw a color swatch, we draw a rectangle around it. Store
each rectangle (64 total) and when we get an onTouch event, test it
against each of the 64 rectangles to see what we hit. It's not
necessarily fast but it's arguably fast enough.

This is a first pass, we just determine what color was hit. The next
pass will return the color to the parent Activity.