PaintBox: Refactor to hold a Context
PaintBox needs a Context to be able to access application
resources. So refactor it to take a Context on creation,
and update everything that creates one to also pass a
Context.

Update WatchFacePreset so it doesn't know about colors, but
only about 6-bit color indexes. Anything that needs an
actual color, go query the PaintBox. Refactor everything
that does that.

Update StateObject to have a PaintBox, and everything can
access their PaintBox through the StateObject.