WatchFaceState: transfer complications code
I decided that the complications code should be in WatchFaceState as
they really affect the model.

WatchFaceGlobalDrawable ought to be concerned with drawing and
presenting the model, not being the model.

Long-term plans for WatchFaceState are that the fields in the object
should be private, and encapsulated in higher-level methods where the
logic lives.