AnalogComplicationConfigActivity: integrate WatchFacePreset
First pass at integrating WatchFacePresetSelectionActivity into the
main codebase. This commit is just enough code to show new selection
options in the main preferences UI for the following:

 * Hour Hand Shape
 * Hour Hand Length
 * Hour Hand Thickness
 * Hour Hand Stalk

None of these actually do anything yet (in fact it causes an exception
if you tap them), it's just putting them in the UI. Next step is to
actually load the current WatchFacePreset and permute them in the
relevant way to give WatchFacePresetSelectionActivity something to
display.