AnalogComplicationConfigData: add "permute"
Add a whole lot of classes that take a WatchFacePreset and permute them
in a variety of ways (i.e. change the HandStalk for the hour hand to
one of everything in the HandStalk enum, etc.)

These live in the "permute" function. We call that and get a String[]
of permutations which we pass into the WatchFacePresetSelectionActivity
which then displays them!

Also add "watch_face_preset_config_list_item" to house the ImageView
that actually holds the image.