AnalogComplicationConfigRecyclerViewAdapter: add bind
Instead of lots of accessors for ViewHolders when we bind them, just
add a "bind" method which takes a ConfigItem, and put the logic in
there. Which simplifies things.

Also make it so we call "permute" just in time, that is, just before
we need to use the results to display some permutations. This means
that if we change the WatchFacePreset it's picked up and we don't use
an old stale copy. The old version just used to get the WatchFacePreset
on startup and use that to permute for everything on startup, which is
inefficient.