WatchPartTicksRingsDrawable: alternate tick styles
Implement the DOT, TRIANGLE and DIAMOND tick styles for drawing in
WatchPartTicksRingsDrawable. Currently they honour the TickThickness
and TickLength settings, so the next steps is to hook up
TickRadiusPosition too (figuring out how that's different for
Four, Twelve and Sixty ticks, if that's worthwhile).

These draw in a new way: we draw each tick at the twelve o'clock
position and then rotate it to where it needs to be. Next steps are to
retrofit BAR to do this too, and probably also the notification dot.