WatchPartDrawable: have height and width
Be a little more rigorous with how we do height and width.

Previously we just used the canvas height and width. Now we act more
like a Drawable and set our own bounds.

This could probably do with a bit more fixing, especially getting rid
of "onWidthAndHeightChanged" and override the native "setBounds".