WatchPartTicksRingsDrawable: split three ways
Make WatchPartTicksRingsDrawable abstract and define concrete classes
for Four, Twelve and Sixty. Yes, this means we'll have to draw three
times. Time will tell whether this was a de-optimisation or not.

This ought to be useful for future use when we highlight the
particular features in settings, and blur/greyscale out the rest.

We'll do something similar with the hour, minute and seconds hands.