WatchPartDrawable: cache unchanging graphics
We previously had caching as part of the ticks and rings layer, but
that was very specific and hard-coded work.

Now, because we're part of a LayeredDrawable, be a bit smarter about
how we cache. It works this way:

The lower layers can be cached. In our case, the background, rings
and ticks can be cached.

When our cache is invalidated and we need to redraw, we do all our
drawing into a cache canvas (rather than the usual display canvas).
When all our cacheable layers have been drawn, the final layer (which
was marked as the "cache point") blits the cache canvas onto the
display canvas. The remaining layers draw as usual.

When our cache is valid, we skip all the redrawing. The "cache point"
layer still blits the cache canvas onto the display canvas.

This apparently works, in that it doesn't crash and it appears to
display correctly. Whether it speeds us up anything is yet to be
determined. I'm also mildly sure I broke night tint and probably also
the notification rings. Put them on the TODO list.