WatchFaceGlobalCacheDrawable: initial commit
Put all our caching logic into WatchFaceGlobalCacheDrawable.

So this is easier yet: any drawables we want cached, put into an
instance of WatchFaceGlobalCacheDrawable. It will do the caching and
hold the cache bitmaps etc.

Our drawables don't need to know any cache logic at all. They can just
draw on their merry way and if it needs to be cached, the parent can
deal with it.