WatchPartDrawable: new bezel drawing concept
Add a new code path for bezel drawing that uses Paths rather than
Strokes from an offscreen Canvas. Also it keeps to the bounds of the
original path rather than exceeding it, so it should look better in
the case that the ticks touch.

In theory this should be faster as it's entirely accelerated. In
practice it seems a decent whack slower.