WatchPartDrawable: fast rendering path
Add a fast rendering path for hands.

We noticed that we spend a lot of time in Path.Op generating the
bezels for our watch hands. Now these watch hands barely move (well,
the second hand does, but not minute or hour). So we only generate the
bezels once every 6 degrees (or once per minute). This is graphically
less precise but who's squinting that hard at their watch anyway?

Also we attempt to cache the paths, by not modifying them. Instead we
keep the hands pointing to 12 o'clock, and rotate the canvas every
time. In theory this is easier on the GPU.