WatchFaceGlobalCacheDrawable: cache active and ambient
Separate caches for active and ambient, so we don't need to recalculate
when switching between active and ambient. That is, when going to
ambient, the recalculation takes a lot of time for something that
shouldn't. This uses up a bit more memory but these watches have 512MB
each, what's another 1MB usage...