WatchFaceGlobalCacheDrawable: improve night vision
Overhaul the night vision tint. Rather than re-drawing the entire
watch face every time the tint changes, just draw everything in white
the first time around, then tint as a post-processing effect. In theory
this is fast as the GPU can do it.

This also has potential to simplify things. We'll see. I did comment
out a whole bunch of code for this.