WatchPartHandsConfigData: initial revision
Initial revision of WatchPartHandsConfig{Data, RecyclerViewAdapter,
Activity}. A very cut-down version of AnalogComplicationConfig classes
with only the code necessary to configure watch part hands.

This code is not yet hooked up to anything, it's just spun out into
its own commit so we can work on it next. We'll probably make this
into some base abstract classes that have common logic and scaffolding.