WatchPartHandsDrawable: hand length overhaul
Refactor our hand lengths.

They are now based on our geometric root-2 series, but we add a
constant 2.0 to them to compress their range a little.

The minute and second hands are now equal length.

The hour hand is (the golden ratio) smaller than the minute hand,
which ought to be aesthetically pleasing if I learned my classics.