WatchPartHandsConfigData: implement it
We figure that any BaseConfigData can be passed in to
AnalogComplicationConfigActivity (which we'll soon rename). Doesn't
have to be the default AnalogComplicationConfigData.

In effect our activity can be re-used with different BaseConfigData
objects to get different effects.

So with some judicious use of intent string extras, we check if
we're asked to use AnalogComplicationConfigData or
WatchPartHandsConfigData when launching an
AnalogComplicationConfigActivity, then set it up accordingly.