Settings: initial revision
Create a new Settings class to hold per-slot settings (i.e. everything
that's not in WatchFacePreset).

Relocate "Unread Notifications" and "Night Vision" booleans into there.

This can do with a bigger refactor later as we can just have a general
toggle control that twiddles the knobs in Settings as needed. This is
just to get us started.