ConfigData: initial WatchFaceDrawableConfigItem
Introducing WatchFaceDrawableConfigItem, a ConfigItem that displays
a WatchFaceGlobalDrawable.

Essentially a live version of the watch face as a ConfigItem.

Currently we have it set up to select complications (and not even
select; in this initial version all it does is display them, and I'm
not sure that works properly either!).

Future versions will be more generic. For example this code is a direct
copy of that in WatchFaceSelectionActivity. We should probably have a
common ancestor.