WatchPartHandsDrawable: better cutout drawing
Get cutouts working with our new inset code. Also get cutouts working
when we don't draw a stalk, in which case we just draw the relevant
half of the cutout. So now we have the concept of a full hand cutout,
a top half hand cutout and a bottom half hand cutout. Calculate all of
these every time, regardless of the type. Assume it's quick.