PaintBox: cache GradientPaint bitmaps
Creating GradientPaint bitmaps for the brushed effect seems like a lot
of effort. Cache and re-use them if we can.

Every time we create a new one we add it to the static "mBitmapCache"
under a WeakReference. So it may be garbage collected (or put another
way, we re-use objects that were already in memory, but we won't cause
memory leaks by trying to hang on to them).

When we need a brushed effect bitmap, we check if there's one in memory
that hasn't been garbage collected; that's what we use.