PaintBox: make weave better
Instead of trying to do transfer modes, just draw everything onto a
temp bitmap (and canvas) and then composite it onto the final bitmap.
Uses more memory for the bitmap, but hey, it simplifies the code. One
day I'll optimise it if people can't spare the extra 500KiB or so...