ConfigData: revert generics changes
Reverts the changes made in:

    2fd8318 SettingsConfigData: add UI for dev mode
    a47a259 ConfigData: make "PickerConfigItem" generic
    57cd625 ConfigData: make "ToggleConfigItem" generic

Bringing the tree back to the state it was in:

    b3df3c5 BytePackable: implement Cloneable

Except for "strings.xml" which contains labels we'll use later on.

We need to redesign. Implementing generic types that work on
descendants of "BytePackable" is probably making things more complex
than necessary for no payoff.