WatchFaceState: move functionality here
Make the basic unit of configurability the WatchFaceState. Was
previously the WatchFacePreset and Settings classes. These are now
closer to pure state objects, with logic in WatchFaceState.

This huge commit is the smallest commit that I could do that would
still build and wouldn't be convoluted... it mostly works; there are
some defects with toggle switches and probably other things I haven't
discovered. Yet.