ConfigData: allow toggles to be hidden
Add a ConfigItemVisibilityCalculator to "ToggleConfigItem" so we can
have logic that toggles can be hidden, in the same way pickers can.

Add logic to "ToggleViewHolder" to query the visible or hidden state
and show or hide as required.

Finally, make "Stats" and "Stats Detail" only visible when developer
mode is active.