Project: multiple slots, part II
Make "ProWatchFaceService" abstract and move everything to slots A, B,
C and D.

Modify "ConfigActivity" and "WatchFaceSelectionActivity" to require
(and work on) individual slots.

At the moment the only independent slot settings is complications.
WatchFacePreset and Settings are still shared, as SharedPref is still
common amongst all slots. That's a fix for a future commit.