WatchFacePreset: implement digits
Implement digits in WatchFacePreset. In the process we bit the bullet
and defined version 0 using a cheap hack, and rearranged everything
in pack/unpack to suit. I think we're using 126 out of 127 bits now...

Removed second hand shape and hard-code it to STRAIGHT.

Also implement digit settings.

I think I broke overrides so we'll have to investigate why...