PaintBox: create fewer Paint objects
Reduce the number of Paint objects we create. Just create a few and
reset them when we need them.

Also move them up to the main PaintBox so we can re-use them. They're
not meant to persist, just to be used to paint and then reset for next
time.

Also make the radial gradient taper in a bit smoother for effect!