WatchFaceState: get labels of linked styles
We want to get the labels of everything with the linked style.

So for "accent highlight" return the label of everything that's been
set to "accent highlight". This gives context for our styles.

It's kind of hard to describe in words and I'm tired and this code has
been hacked together and could do with some design.

Also it doesn't work. But it's probably a start...