BytePackable: limit Style to gradients
Style used to be flat colors (FILL, ACCENT, HIGHLIGHT, BASE) as well as
gradients (FILL_HIGHLIGHT, ACCENT_FILL, ACCENT_HIGHLIGHT, BASE_ACCENT).

Make it only be gradients now.

Those gradients can also be flat colors with the appropriate setting
(i.e. StyleGradient.FLAT). So we're not losing much.

We do save 1 bit for each Style we encode (so 7 bits in
WatchFacePreset) plus another 3 bits in WatchFacePreset as we don't
encode the background style any more (it's hard-coded to
Style.BASE_ACCENT now; if you want to change the background, change
that style).

Update the Settings bitfield to version 3 and WatchFacePreset to
version 2.