PaintBox: slower triangle gradient
We use a slower triangle gradient implementation which draws the
gradient pattern in grayscale [0..255], then maps that to a CIE LAB
gradient in a color look-up-table ("cLUT").

This is more accurate and better-looking than the equivalent sRGB.

The mapping takes time (75ms, 5 or so frames?) which is an eternity,
because we iterate over every pixel in the bitmap when mapping from
grayscale to color, which takes time. We use a Java lambda for the
mapping which in principle is a speedy way to do things as it uses
smarts like multi-threading etc. In practice a careful C implementation
in the NDK would probably trounce it. Add that to the TODO list.