PaintBox: faster addTriangleGradientSlow
Rather than do a naive stream of the array into a mapping, requiring
an entire bitmap-sized array and another copy of it for the output...

Go line by line, and unroll the loop to do 8 pixels at a time. This
only uses an array the size of one row, so less GC as a bonus.

I did a few alternate designs and this one benchmarks to be the
fastest, which doesn't say much; it's still about ~80ms on a good day.

About half of that however is the call to getPixels, which means that
it's rendering in software then writing it out. So if I do a native
implementation it'll still have that overhead.