WatchPartBackgroundDrawable: draw full canvas
When drawing the background, we now have options to draw the entire
canvas, or just a part that's shaped according to the hardware watch.

We don't actually determine what shape the hardware watch is (i.e. we
don't yet have logic to detect round, round-with-cutout or square).
Instead we go whimsical and draw a dodecagon (a.k.a. 12-gon).