WatchPartRingsDrawable: implement overlap
The way I've done it, overlap is just the rings, not the complication
drawable within; it's more akin to padding on the complication.

Not sure if I like it this way and it's not what I intended (rather the
complication drawable should scale bigger/smaller) but this could be
serendipity and I may or may not keep it!