WatchFaceState: implement complication size
Using lots of formulae from Wikipedia and a misspent youth doing lots
of triangle and circle geometry using terms such as "incircle".

Our previous code hard-coded circle sizes based on 6 complications or
so. This code scales the sizes exactly to the number of complications
and further on the user-selected size.