WatchPartDrawable: inner glow initial commit
Initial revision of inner glow.

Not convinced this is the right way to do things, but right now it's
gated behind developer mode stats, so let's leave the framework there.

I don't think drop shadows work in a proper way in hardware rendering,
and this is dependent on API level 28 too. Revisit this later...