BaseRecyclerViewAdapter: hardware backed ImageView
Set the layer type for our ImageViews to View.LAYER_TYPE_HARDWARE.

This is one of those "one line of code: $1; knowing which line of code
to write: $999" scenarios.

Our slow performance in scrolling our ImageViews is due to having to
redraw the entire thing on the GPU each frame. The GPU is fast but
it's still not an efficient way of doing things.

When we set View.LAYER_TYPE_HARDWARE, we make an implicit promise: we
won't update the Drawable any more than necessary, and you can assume
it's frozen. So rather than render it every frame, you can render it to
a texture and use that.

It's more expensive up-front, and requires assumptions that don't
always hold; that's why it's not the default. But if you know the
option is there, the rewards are great.