WatchFaceState: add hand cutout toggle
Change our state object and settings: now we have a "Hand Cutout"
toggle. When off, use "HandCutoutCombination.NONE". When on, show
the settings to modify "Hand Cutout Shape" and "Hand Cutout Style".

When toggling off, we save the current HandCutoutCombination to
Settings (as "previous") so we can restore it when we toggle on again.