SharedPref: 5-minute limit for new history items
When we get a new watch face string, only store the old one in history
if we haven't stored anything in the last 5 minutes.

This means that if we make a raft of changes over 5 minutes, only the
most recent gets added to history.

Also re-use some objects rather than creating them fresh every time.