ComplicationHolder: remove unused caching code
There was a bunch of caching code for ComplicationHolder that I once
wrote. Plus a lot of associated code to do with invalidating Drawables,
which I imagine was used to enable caching, since removing it doesn't
seem to have any negative impact.

The code hasn't been used in a while and it's unclear it would speed
things up anyway (at least, things seem plenty fast without it). And it
just makes things more complex.

So remove it. If I want it back, I can probably re-code it from scratch
with whatever future skills I have.