BaseRecyclerViewAdapter: show icon if no permissions
If we need permissions but don't have them, don't draw the color
swatch, draw an icon instead. And don't show the color name.

And use ConfigActivity.UPDATED_CONFIG_REDRAW_NO_MATTER_WHAT_RESULT_CODE
so that we always redraw, so on the off-chance that the user grants
permissions, we can get rid of the icon and draw the swatch.