PaintBox: prepare to draw bitmaps
When we finish our drawing into Bitmap objects, we call
"prepareToDraw", which hints to Android that we won't be making any
more changes to the Bitmap and it can start copying it to the GPU
right away rather than at the last moment.

I'm not sure there was enough of a bottleneck that this will help much,
but it may have an effect in congested situations. It's good practice
too.

Also add a missing copyright notice to "WatchFaceGlobalCacheDrawable".