WatchPartDrawable: improved bezels drawing
Improve the logic behind WatchPartRingsDrawable, which was actually
quite terrible. Make the rings thicker (was 4% of the complication
size, now 10%), which fixes the visual glitch in issue #1.

Improve the bezel drawing. This is a first pass with some duplicated
code still to cull out, but the concepts are there. Instead of drawing
bezels on two sides (northwest and southeast with our bezel paints) we
draw on four sides (the above, plus also northeast and southwest with
a translucent white and translucent black paint).

This adds a bit of light-and-shadow definition to the bezels which
makes them pop a bit more.

Also commonise the code behind the "fast-drawing", which pre-caches
the bezels for commonly rotating paths (i.e. the hands), as generating
paths for the bezels is potentially quite expensive.