Project: rename "Style" to "Material"
For every instance of "Style" where we mean it as "a combination of a
colour and a texture", rename that to "Material". It's less ambiguous
and perhaps a bit more evocative of how we build a watch face and where
we want to go in the future.

Affects both code and UI (resources).

"Text Styles" stay as they are, as do every current instance of Style
from the android.graphics namespace, etc.