ConfigData: rewrite Mutator
Rewrite our Mutator logic code to be less of a gross hack, and more of
a refined hack.

Instead of returning a String[] of WatchFaceState configurations, and
a list of names, pack both into a Permutation class. Our new Mutator is
only one method and can be called as a lambda expression.

Remove the requirement to provide code to check whether a given
configuration is active; we'll calculate that in the caller. This also
means we don't need most of those "getters".

And a whole bunch of supporting code, some of which isn't hooked up yet
and is commented out. This will be pressed into service in upcoming
commits!