WatchPartDrawable: change active exclusion path
Change the default active exclusion path to cover the entire screen,
whether round or rectangular.

In practice this means that the pip background "ring", if we have one,
now won't be drawn with an outer circular bezel; it will extend to the
edge of the screen. It's 50/50 whether this is an improvement, however
if it's dark then it'll blend better with the screen edge.