WatchFaceSelectionActivity: scroll to correct item
When opening the activity, we scroll to the currently selected item.

We previously used "smoothScrollToPosition" which often had trouble
scrolling to the right item, overshooting etc. Now we prepare by using
"scrollToPosition" to get us into the right area first, then the
"smoothScrollToPosition" to put us into the exact location.