Project: support decomposable rendering
This series of commits adds decomposable watch face rendering, to make
use of the offload processor for ambient display. There are two main
benefits: it allows a ticking seconds hand in ambient mode, and it's
(in theory) more energy efficient.

There are two fairly large drawbacks too: it's added code complexity
(just look at it!) and there are defects with Google's implementation
of watch face offload (sometimes it fails to update and freezes, or
blanks out).

In this commit, we add all the scaffolding and support code for watch
face decompositions:

We use the AlarmManager to wake up Wear OS so we can update our watch
face decomposition as the situation changes. And associated
BroadcastReceivers and Intent code to trigger updates.

We add "isDecomposable" to WatchFaceState to check whether to use our
new code! (Currently hard-coded to "false".)

We add code interfaces to WatchFaceGlobalDrawable to enable drawing
decompositions. The interfaces are stubs right now.

And a bunch of minor code and documentation updates.