PaintBox: add bitmap color reduction
In watch face decompositions, we're limited to a palette of 16 colors
(from the RGB332 palette). However we draw a lot more colors than that.

Fortunately we already have a fast and easy way to convert full-color
images back to 16 colors: we construct a color look-up table (cLUT)
and use our existing RenderScript kernel to map the images with the
cLUT.

Hook up this code.