WatchFaceGlobalCacheDrawable: enable decomposition
Add supporting code to WatchFaceGlobalCacheDrawable to enable it to
render to a watch face decomposition as well.

It renders the "base" image of the decomposition: everything in the
ambient watch face cache (digits and ticks) plus non-time-dependent
complications. (So if a non-time-dependent complication receives an
update, we re-render this component.)