WatchFaceComplicationsDrawable: enable decomposition
Add supporting code to WatchFaceComplicationsDrawable to enable it to
render to a watch face decomposition as well.

It renders all time-dependent complications into a "time-dependent
strip", which is a a NumberComponent and a FontComponent. It's rather
complicated code which replaces the stock Wear OS
ComplicationComponent, but because we have control it looks much
better. (No anti-aliasing.)

Note: does not render any non-time-dependent complications (this is
handled in WatchFaceGlobalCacheDrawable).