Project: add offload mode on/off setting
Add a setting to enable (or disable) offload mode. Add scaffolding
code to persist it.

Hook this setting into the decomposition render code to make it work.

Add a ToggleConfigItem to enable (or disable) it in the settings
activity. It will only show if offload mode is supported on this
device.