PaintBox: new material renderer
A new material renderer based on RenderScript.

It's an expanded version of the triangle renderer, which uses a 256x1
lookup table to map colors onto the material. In this case we use a
64x32 lookup table, which not only maps the gradient hues (on one axis)
but also the lightness and darkness (on the other). Because these are
calculated via CIELUV, it theoretically looks better.

Because we pre-generate all materials and textures (in RenderScript),
the only thing we do at runtime is the RenderScript mapping. So in
principle this solution should be faster too, or at least, consistently
fast across all operations.