WatchFaceGlobalDeferredDrawable: initial revision
Add WatchFaceGlobalDeferredDrawable, a WatchFaceGlobalDrawable variant
which just draws a simple placeholder to begin with, then (in the
background) draws the full watch face, then invalidates and refreshes.

So, a lazy-loading version of WatchFaceGlobalDrawable, which doesn't
freeze the UI when first displaying a watch face.

Now scrolling through a BaseRecyclerViewAdapter is much faster.