WatchPartComplicationsDrawable: only draw active
Only draw active, foreground complications. Usually, drawing an
inactive complication would be a no-op as it's of type
ComplicationData.TYPE_NOT_CONFIGURED, which just skips drawing. However
occasionally it can shift to another "inactive" type such as
ComplicationData.TYPE_NO_DATA, which does draw... and it's either a
placeholder "- -" string, or (for a background image), a black image.

The above caused the everything below the complication layer to be
overdrawn by a black image in the above circumstances; this is the fix.