PaintBox: use the Oklab colorspace
We've used a few colorspaces: we started with sRGB, then converted to
CIELUV for better perceptual rendering. The next step is Oklab, which
is more accurate for lightness, hue and saturation; for example, CIELUV
draws blue gradients that are hue-shifted towards purple, whereas Oklab
draws them blue.

At this point, we use Oklab mainly for gradient rendering; however in
the future we may use it to predict a better six-bit palette too.

Implemented in both Java and RenderScript.

More info, the formulas and the constants are available from:
https://bottosson.github.io/posts/oklab/

Hello world! If you've downloaded the latest version of Watch Kit Pro
and noticed that some watch faces look a bit different, this is why...