ConfigActivity: remove navigation drawer
Look, the WearableNavigationDrawerView was a relic of an older Wear OS.
The modern style (Wear OS 3 onward) doesn't use them. I thought it was
a fine way of presenting a UI, but it's not part of the Google vision,
so now I suppose that neither is it part of ours.

Plus, the drawer had a light-coloured background and that tripped up
the Wear OS quality guideline 'WO-V13', which states that the
background must be black, or you can't publish to the Play Store.
Thanks again, G.